refactor for a cleaner diff
